+++ src/entities/news/components/NewsCard.tsx
