+++ src/entities/news/components/NewsList.tsx
