+++ src/entities/news/components/NewsPaginatedList.tsx
