--- src/shared/components/ui/skeleton.tsx
@@ -1,7 +1,0 @@
-import { cn } from '@/shared/lib/utils';
-
-function Skeleton({ className, ...props }: React.ComponentProps<'div'>) {
-  return <div data-slot="skeleton" className={cn('bg-accent animate-pulse rounded-md', className)} {...props} />;
-}
-
-export { Skeleton };
